refactor execution context models and behaviour loading
- Following metamodel changes in `minhnh/metamodels-bdd#8`, concepts
  previously under simulation module will be moved to execution module
  for shared use in different execution contexts.
- Split behaviour and scenario execution logic into dedicated `execution`
  modules to avoid circular import errors.
- Update execution test and various refs accordingly.
- Also include automated styling changes from ruff.
