+++ src/bdd_dsl/models/environment.py
@@ -3,6 +3,7 @@ from typing import Any, Optional, Generator
 from rdflib import Graph, URIRef
 from rdflib.exceptions import UniquenessError
 from rdf_utils.models.common import AttrLoaderProtocol, ModelBase, ModelLoader, get_node_types
+from bdd_dsl.execution.common import load_attr_has_config
 from bdd_dsl.models.queries import Q_MODELLED_OBJECT
 from bdd_dsl.models.urirefs import (
     URI_ENV_PRED_HAS_OBJ,
@@ -12,7 +13,6 @@ from bdd_dsl.models.urirefs import (
     URI_ENV_TYPE_WS_OBJ,
     URI_ENV_TYPE_WS_WS,
 )
-from bdd_dsl.simulation.common import load_attr_has_config
 
 
 class ObjectModel(ModelBase):
@@ -29,14 +29,14 @@ class ObjectModel(ModelBase):
         load_attr_has_config(graph=graph, model=self)
 
         for model_id in graph.objects(subject=obj_id, predicate=URI_ENV_PRED_HAS_OBJ_MODEL):
-            assert isinstance(
-                model_id, URIRef
-            ), f"unexpected type for model ID (not URIRef): {type(model_id)}"
+            assert isinstance(model_id, URIRef), (
+                f"unexpected type for model ID (not URIRef): {type(model_id)}"
+            )
 
             obj_model = ModelBase(graph=graph, node_id=model_id)
-            assert (
-                obj_model.id not in self.models
-            ), f"Obj '{self.id}' has duplicate models '{obj_model.id}'"
+            assert obj_model.id not in self.models, (
+                f"Obj '{self.id}' has duplicate models '{obj_model.id}'"
+            )
             self.models[obj_model.id] = obj_model
 
             for model_type in obj_model.types:
@@ -81,9 +81,9 @@ def _get_ws_objects_re(
     assert ws_id in ws_dict, f"_get_ws_objects_re: '{ws_id}' not on record"
     ws_model = ws_dict[ws_id]
     for sub_ws_id in ws_model.workspaces:
-        assert (
-            sub_ws_id in ws_dict
-        ), f"sub-ws '{sub_ws_id}' of workspace '{ws_model.id}' not on record"
+        assert sub_ws_id in ws_dict, (
+            f"sub-ws '{sub_ws_id}' of workspace '{ws_model.id}' not on record"
+        )
         for obj_id in _get_ws_objects_re(ws_id=sub_ws_id, ws_dict=ws_dict, ws_path=ws_path):
             yield obj_id
 
@@ -112,24 +112,24 @@ def _load_ws_re(
         ws_comp_id = graph.value(object=ws_id, predicate=URI_ENV_PRED_OF_WS, any=False)
     except UniquenessError as e:
         raise RuntimeError(f"multiple composition for workspace '{ws_id}': {e}")
-    assert isinstance(
-        ws_comp_id, URIRef
-    ), f"no ref to composition URI for ws '{ws_id}': {ws_comp_id}"
+    assert isinstance(ws_comp_id, URIRef), (
+        f"no ref to composition URI for ws '{ws_id}': {ws_comp_id}"
+    )
 
     ws_comp_types = get_node_types(node_id=ws_comp_id, graph=graph)
 
     if URI_ENV_TYPE_WS_OBJ in ws_comp_types:
         for obj_id in graph.objects(subject=ws_comp_id, predicate=URI_ENV_PRED_HAS_OBJ):
-            assert isinstance(
-                obj_id, URIRef
-            ), f"comp '{ws_comp_id}' of ws '{ws_id}' does not ref an obj URI: {obj_id}"
+            assert isinstance(obj_id, URIRef), (
+                f"comp '{ws_comp_id}' of ws '{ws_id}' does not ref an obj URI: {obj_id}"
+            )
             ws_model.objects.add(obj_id)
 
     if URI_ENV_TYPE_WS_WS in ws_comp_types:
         for sub_ws_id in graph.objects(subject=ws_comp_id, predicate=URI_ENV_PRED_HAS_WS):
-            assert isinstance(
-                sub_ws_id, URIRef
-            ), f"'{ws_comp_id}' of ws '{ws_id}' does not ref a ws URI: {sub_ws_id}"
+            assert isinstance(sub_ws_id, URIRef), (
+                f"'{ws_comp_id}' of ws '{ws_id}' does not ref a ws URI: {sub_ws_id}"
+            )
 
             ws_model.workspaces.add(sub_ws_id)
             _load_ws_re(ws_id=sub_ws_id, graph=graph, ws_dict=ws_dict, ws_path=ws_path)
@@ -149,9 +149,9 @@ class EnvModelLoader(object):
 
     def _query_obj_models(self, graph: Graph) -> Graph:
         q_result = graph.query(Q_MODELLED_OBJECT)
-        assert (
-            q_result.type == "CONSTRUCT" and q_result.graph is not None
-        ), "querying ModelledObject's failed"
+        assert q_result.type == "CONSTRUCT" and q_result.graph is not None, (
+            "querying ModelledObject's failed"
+        )
         assert len(q_result.graph) > 0, "querying for ModelledObject's returned an empty graph"
 
         return q_result.graph
