+++ src/bdd_dsl/models/uri.py
@@ -8,6 +8,7 @@ URI_TRANS = f"{URL_SECORO}/transformations/"
 URI_MM_TASK = f"{URL_SECORO_MM}/task#"
 URI_MM_AGENT = f"{URL_SECORO_MM}/agent#"
 URI_MM_BHV = f"{URL_SECORO_MM}/behaviour#"
+URI_MM_EXEC = f"{URL_SECORO_MM}/execution-context#"
 URI_MM_SIM = f"{URL_SECORO_MM}/simulation#"
 URI_MM_OBS = f"{URL_SECORO_MM}/observation#"
 URI_MM_BDD = f"{URL_SECORO_MM}/acceptance-criteria/bdd#"
